AMD64HotSpot: emit jump to IC miss handler directly instead of emitting it inside verfied_entry
for empty methods we emitted something like this if a inline cache is needed:

    prefix:
        00:    < IC check >
               ...
        0b:    jne <ic_miss_call>
        11:    nop
               ...
    verified_entry:
        20:    ret
    ic_miss_call:
        21:    jmp <entry of runtime function>


when a method is deoptimized, HotSpot patches the verified_entry (0x20) with a
jump to a stub that handles call-sites that has been made non-entrant. since this jump
is 5 bytes long, it will overwrite ic_miss_call and blow up every caller that
calls this method via the unverified entry (prefix).

the fix is to emit the jump to the runtime function inside the unverfied entry:

    prefix:
        00:    < IC check >
               ...
        0b:    je  <verified_entry>
        11:    jeq <entry of runtime function>
        16:    nop
               ...
    verified_entry:
        20:    ret

--HG--
extra : rebase_source : f189d42bb5dbe59ce74c227a19505f98b1192633