TruffleDSL: generate polymorphic `createUninitialized()` method on Node classes.
This method is useful to get an uninitialized copy of an AST node.
It uses the copy constructor for the uninitialized state and passes `this` to it.
In TruffleSOM it is used to copy an AST for inlining. This way, there is no need to keep copies of the original ASTs around.

Signed-off-by: Stefan Marr <git@stefan-marr.de>